[media] media: media controller entity framework enhancements for ALSA
Add a new media entity operation register_notify is added
to media_entity_operations structure. This hook is called
from media_device_register_entity() whenever a new entity
is registered to notify other entities attached to that
media device of the newly created entity. Entity owners
can register the hook to create links and take any other
action when a new entity is registered. A new field is
added to the struct media_entity for entity owners to save
their private data that is used in the register_notify hook.

Signed-off-by: Shuah Khan <shuahkh@osg.samsung.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>