media: rkisp1: Add user space ABI definitions
Add the header for userspace

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Signed-off-by: Jacob Chen <jacob2.chen@rock-chips.com>
[update for upstream]
Signed-off-by: Helen Koike <helen.koike@collabora.com>

Series-changes: 7
- Fix checkpatch errors (lines over 80 and SPDX)
- Add TODO to improve docs

Commit-notes:
Hi,

I don't have the "REF_01 - ISP_user_manual, Rev 2.57" document that was
mentioned in previous version, so I'm adding a TODO for now to improve
the docs once we collect the information.
If Rockchip people could help here it would be great.

Thanks
Helen
END