rkisp1: rockchip-dphy: A few minor fixes
* Normalize struct names to rk_dphy.
* Simplify some names.
* Remove unused stuff.
* Move checks to configure.

I wonder if it's OK to call this driver like this, vs. CSI-2 or
something (it doesn't seem like we'll merge DSI support here)?

PS: Remember to discuss the memcpy vs. struct assignment that
Laurent pointed out. Some people dislike the assignment because
it hides the fact that there is a memcpy happening. OTOH,
the assignment is safer because it checks types.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>