rkisp1: try/set rework
The main reason for this is that bool arguments
should be avoided generally speaking, although
there might be exceptions. OTOH, it's fine to return bools.

Examples:

  bool do_something(bool try);

This will look cool in the prototype, but it won't
look cool while using it:

  done = do_something(true);

It's not obvious what's the 'true' argument for,
only the 'done' variable makes some sense.

This is better:

  bool try_something(void);
  bool do_something(void);

And when we use it, everything is obvious,
the API barely needs any doc.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>