This commit is only to test the following hypothesis:
In rksip1-isp.c, phy_mipi_dphy_get_default_config is called as

       phy_mipi_dphy_get_default_config(pixel_clock, isp->sink_fmt->bus_width,
                                         sensor->lanes, cfg);

of course bus_width is set to 8

However in phy_mipi_dphy_get_default_config the second parameter is
interpreted as "bpp" (bits per pixel).
So it should be 16!

Of course that modification should work only for ADV7280 and will break things
for OV5647 and others.
It should be reimplemented in the right way when the hypothesis is confirmed.